EasyOcr/requirements.txt
fastapi
uvicorn[standard]
paddleocr
paddlepaddle
opencv-python-headless
numpy<2
Pillow
PyMuPDF
python-multipart
